Move tests to module in own file
